fix(types): correct length() and blob function return types for SQLite compatibility
- length() now returns character count instead of byte count for UTF-8 strings
  (e.g., length(char(350,351,352,353,354)) returns 5, not 10)
- randomblob(), unhex(), zeroblob() now return SqlValue::Blob instead of Varchar
  (typeof(randomblob(32)) now correctly returns "blob")
- Updated tests to match new behavior
- Added "main" schema support in catalog error tests

Refs #4607

+++ crates/vibesql-executor/tests/string_tests/length_functions.rs
@@ -192,7 +192,8 @@ fn test_length_basic() {
 #[test]
 fn test_length_multibyte() {
     let (evaluator, row) = create_test_evaluator();
-    // LENGTH returns byte count (unlike CHAR_LENGTH)
+    // SQLite: LENGTH() returns character count, not byte count
+    // Use OCTET_LENGTH() for byte count
     let expr = vibesql_ast::Expression::Function {
         name: vibesql_ast::FunctionIdentifier::new("LENGTH"),
         args: vec![vibesql_ast::Expression::Literal(vibesql_types::SqlValue::Varchar(
@@ -201,8 +202,8 @@ fn test_length_multibyte() {
         character_unit: None,
     };
     let result = evaluator.eval(&expr, &row).unwrap();
-    // "café" is 5 bytes in UTF-8
-    assert_eq!(result, vibesql_types::SqlValue::Integer(5));
+    // "café" has 4 characters (é is one character, 2 bytes in UTF-8)
+    assert_eq!(result, vibesql_types::SqlValue::Integer(4));
 }
 
 #[test]
